End of day 6/24
+ Player states placed in a separate coroutine for execution efficiency.
+ Player movement tweaked specifically how dodging and running work.
+ Shooting, hurting enemies, and getting hurt fixed
+ Interaction notification fixed

Still Broken:
For reasons unknown it seems you can still set the dodging,
knockback(kicking), and interacting flags even during the animation
execution despite the animComplete check.

As discussed I believe it may be easier to instead of basically guessing
at the timing and manually setting an animation complete bool when
animations finish, we can make use of Unity's animation classes to get
this info. This will need to be investigated and played with further.