Work 7/22
Fixed the ShirtSwitch script to crudely modify the mesh the player is
wearing based on public meshes applied to the script as reference. All
you need to do is change the public ID in the script to have the armor
change while playing. May be a sloppy way of doing it but it works
without deleting and screwing with any meshes. Note that "chest0" is an
empty mesh and is used when the player is top less.

I will be working on a new model that matches the game's theme a bit
more, but at least with this we can see that the armor swapping and
animation imports from Blender works. Still gotta fix that wonky run
animation, but that will come with time.